+++ psalm.xml
@@ -52,11 +52,10 @@
 
         <InternalMethod>
             <errorLevel type="suppress">
-                <file name="src/Bundle/Controller/ParametersParser.php" />
-                <file name="src/Bundle/Controller/RequestConfiguration.php" />
                 <file name="src/Bundle/Controller/ResourceController.php" />
                 <file name="src/Bundle/Grid/Parser/OptionsParser.php" />
                 <file name="src/Component/Symfony/Request/ParametersParser.php" />
+                <file name="src/Component/Symfony/Request/RequestConfiguration.php" />
             </errorLevel>
         </InternalMethod>
