+++ src/Bundle/Controller/AuthorizationCheckerInterface.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\Controller;
 
+use Sylius\Component\Resource\Symfony\Request\RequestConfiguration;
+
 interface AuthorizationCheckerInterface
 {
     /**
