+++ src/Bundle/spec/Controller/DisabledAuthorizationCheckerSpec.php
@@ -15,7 +15,7 @@ namespace spec\Sylius\Bundle\ResourceBundle\Controller;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Bundle\ResourceBundle\Controller\AuthorizationCheckerInterface;
-use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
+use Sylius\Component\Resource\Symfony\Request\RequestConfiguration;
 
 final class DisabledAuthorizationCheckerSpec extends ObjectBehavior
 {
