+++ src/Bundle/spec/Controller/ResourcesResolverSpec.php
@@ -15,10 +15,10 @@ namespace spec\Sylius\Bundle\ResourceBundle\Controller;
 
 use Pagerfanta\Pagerfanta;
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
 use Sylius\Bundle\ResourceBundle\Controller\ResourcesResolverInterface;
 use Sylius\Component\Resource\Model\ResourceInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
+use Sylius\Component\Resource\Symfony\Request\RequestConfiguration;
 
 final class ResourcesResolverSpec extends ObjectBehavior
 {
