+++ src/deprecation.php
@@ -32,6 +32,7 @@ foreach ($deprecatedInterfaces as $oldInterfaceName => $interfaceName) {
 
 $deprecatedClassesWithAliases = [
     Sylius\Bundle\ResourceBundle\Controller\ParametersParser::class => Sylius\Component\Resource\Symfony\Request\ParametersParser::class,
+    Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration::class => Sylius\Component\Resource\Symfony\Request\RequestConfiguration::class,
 
     Sylius\Bundle\ResourceBundle\Provider\RequestParameterProvider::class => Sylius\Component\Resource\Symfony\Request\RequestParameterProvider::class,
 ];
